Potion creation
After multiple hours of debugging, I finally was able to create and add the potion based on a known recipe to the player's inventory. For whatever reason, a UE_LOG fixed a bug, which should not be possible. When trying to add the potion to inventory, it couldn't enter into the for loop, for whatever reason. When putting a log to check the number of inventory slots, it started working. SMH! Again, this shouldn't be possible.