Inventory and potion qualities
Implemented different potion qualities. Each potion needs to decide how to handle the different qualities. The test potion heals more the better the quality. Each potion has a hash code, which determines its position in the inventory. If 2 potions are the same, then they will be put in the same slot.