EQS with gameplay tags C++
Gameplay tag test didn't work normally and others were having the same problem, didn't find a solution, so I had to make it in C++.

Added a CustomAIContainer class that holds all the actors that the EnvQuery result gives. After the query, the HandleQueryRequest finds all actors that have matching Gameplay Tags and adds them to the CustomAIContainer object.