Migrated level AGAIN
For whatever reason, when creating a new branch from the master branch that should have had the same data that the AI branch had because I merged the two branches together beforehand, the new branch created did not receive the updates that the merge was supposed to have. SO, I had to migrate the map that I made inside the Level Branch to another project, then deleted the branch because it was a mess because of all the merge issues, and now migrated the level back. What a fuck fiesta. :+1: