More RoadSpline stuff
Dabbled a bit in Component Visualizers and came to the conclusion that it's not worth the effort for the time being and settled for a simpler solution. Added a blueprint function that finds the closest  spline point to the actor selected. This will be useful for long distance AI navigation and just the beginning.