Destination reached logic
Can add a NavDestination actor to the AI. AI will move to the closest spline point to the NavDestination actor. If not set, AI will move to the end of the last road. When Destination is reached, the AI will be set to the AI state that is selected in the BTTask_GetNextSplinePoint.