Navigation works probably
System will calculate the fastest* route to the navigation destination using the road network. AI moves to the NavDestination using roads correctly. *by fastest, it calculates the route that uses the least number of roads, not necessarily the fastest. This needs to be changed in the future and add heuristics to the mix.