EQS fixes
Fixed the EQ_GetHidingSpotsNearStimulus query. Removed the cone generator and replaced it with a grid. Re-did the whole thing and it started working, even though I'm pretty sure that everything was the same as it used to. Debugged this stuff for several hours. Optimized normal HidingSpots query. Average runtime went from 3.5 ms to  less than 1.5 ms.