Dynamic navlinks
Added dynamic nav link proxies. This means that when the player jumps it records creates a nav link proxy and when the player lands it connects the nav points so that AI can navigate in a similar way to the player. Currently buggy and feature incomplete, but the basic feature works as intended (sometimes).