Cover
Added an actor class that implements the IGameplayTagAssetInterface so that they can be used in Env queries. Added gameplay tags for different types of cover. Added env queries that find cover close and another query for finding a good position near the cover. Added Behavior tree tasks that find cover and move to cover. I really need to implement a better way of checking if character is visible. This should be done by tracing for different bones. This should also be an async task.