WorkingMemory and AIGoal
Added WorkingMemory class since it seems to be an important part of the F.E.A.R. AI. Added a few member variables and functions to the AIGoal class. Most Importantly, completed the original FEAR game on the highest difficulty. Had to buy the game just so that the source code will make more sense, and it does.